+++ config/src/main/java/com/alibaba/nacos/config/server/service/trace/ConfigTraceService.java
@@ -19,8 +19,8 @@ package com.alibaba.nacos.config.server.service.trace;
 import com.alibaba.nacos.common.utils.MD5Utils;
 import com.alibaba.nacos.config.server.constant.Constants;
 import com.alibaba.nacos.config.server.utils.LogUtil;
-import com.alibaba.nacos.manager.ConfigMetricsConstant;
-import com.alibaba.nacos.manager.MetricsManager;
+import com.alibaba.nacos.metrics.manager.ConfigMetricsConstant;
+import com.alibaba.nacos.metrics.manager.MetricsManager;
 import com.alibaba.nacos.sys.utils.InetUtils;
 import com.alibaba.nacos.common.utils.StringUtils;
 import org.springframework.stereotype.Service;
