+++ metrics/src/main/java/com/alibaba/nacos/metrics/manager/MetricsManager.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.manager;
+package com.alibaba.nacos.metrics.manager;
 
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.Gauge;
